// ui/client/types.gen.ts
export type PostResponse = {
  /**
   * A URL to the JSON Schema for this object.
   */
  readonly $schema?: string;
  author?: string;
  content?: string;
  created_at: string;
  description?: string;
  feed_id: number;
  guid: string;
  id: number;
  is_read: boolean;
  link: string;
  published_at?: string;
  title: string;
  updated_at: string;
};